Fixed all issues with flash

+++ inc/encoders.h
@@ -16,10 +16,10 @@
 typedef struct{
     uint32_t value;
     uint32_t old_value;
-    uint16_t step;
-    uint16_t min;
-    uint16_t max;
-    uint16_t coef;
+    uint32_t step;
+    uint32_t min;
+    uint32_t max;
+    uint32_t coef;
     uint32_t pin1;
     uint32_t pin2;
     uint8_t old_state;
